Feature/context refactor (#36)
* refactor contexts

* flip loggedin bool
