+++ frontend/src/main.tsx
@@ -1,23 +1,17 @@
-import React, { createContext, useContext, useState } from "react";
+import React, { useContext, useState } from "react";
 import ReactDOM from "react-dom/client";
 
-import { BrowserRouter as Router, Routes, Route, Navigate } from "react-router-dom";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Navigate,
+} from "react-router-dom";
 import * as Pages from "./pages";
 import Layout from "./components/Layout";
+import { AuthContext } from "./contexts/auth.ts";
 
 import "./global.css";
-
-interface AuthContextProps {
-  isLoggedIn: boolean;
-  login: () => void;
-}
-
-// Handle Auth State- Vulnerable to XSS?
-export const AuthContext: React.Context<AuthContextProps> = createContext<AuthContextProps>({
-  isLoggedIn: false,
-  login: () => {},
-});
- 
 
 //If not logged in, route to login
 const PrivateRoute = ({ element }: { element: React.JSX.Element }) => {
@@ -25,32 +19,32 @@ const PrivateRoute = ({ element }: { element: React.JSX.Element }) => {
   return isLoggedIn ? element : <Navigate to="/" />;
 };
 
-
-
-export function App(): React.JSX.Element {
+export default function App(): React.JSX.Element {
   //Handle loggedin state
   const [isLoggedIn, setIsLoggedIn] = useState(false);
-  
+
   const login = () => {
     setIsLoggedIn(true);
-  }
+  };
 
-  
   return (
-    <AuthContext.Provider value={{isLoggedIn, login}}>
-    <Router>
-      <Routes>
-      <Route path="" element={<Pages.Login />}/>
-      <Route path="oauth/callback" element={<Pages.Callback />} />
-        <Route path="/app/" element={<PrivateRoute element={<Layout />} />}>
+    <AuthContext.Provider value={{ isLoggedIn, login }}>
+      <Router>
+        <Routes>
+          <Route path="" element={<Pages.Login />} />
+          <Route path="oauth/callback" element={<Pages.Callback />} />
+          <Route path="/app/" element={<PrivateRoute element={<Layout />} />}>
             <Route path="assignments" element={<Pages.Assignments />} />
-            <Route path="assignments/:id" element={<Pages.AssignmentDetails />} />
-            <Route path="grading" element={<Pages.Grading />}/>
+            <Route
+              path="assignments/:id"
+              element={<Pages.AssignmentDetails />}
+            />
+            <Route path="grading" element={<Pages.Grading />} />
             <Route path="settings" element={<Pages.Settings />} />
-            <Route path="dashboard" element={<Pages.Dashboard />}/>
-        </Route>
-      </Routes>
-    </Router>
+            <Route path="dashboard" element={<Pages.Dashboard />} />
+          </Route>
+        </Routes>
+      </Router>
     </AuthContext.Provider>
   );
 }
